added custom hook for ElementCreator, and storing elements in local storage
