+++ client/src/components/ElementCreator.js
@@ -1,12 +1,8 @@
-import React, { useState, useEffect } from 'react'
+import React, {useState} from 'react'
+import {useElementCreator} from '../hooks/useElementCreator'
 
 const ElementCreator = () => {
-    const [elements, setElements] = useState([
-        {
-            id: 1,
-            name: 'div',
-        }
-    ])
+    const [elements, setElements] = useElementCreator()
 
     const [newElement, setNewElement] = useState('')
 
@@ -21,24 +17,8 @@ const ElementCreator = () => {
             id: nextId,
             name: newElement
         }])
-        console.log(elements)
-        localStorage.setItem('elements', elements)
+        localStorage.setItem('elements', JSON.stringify(elements))
     }
-
-    useEffect(() => {
-        const body = document.body
-        for(let i = 0; i < elements.length; i++) {
-            let element = elements[i]
-            let hasChild = document.getElementById(`${element.id}`)
-            if (!hasChild) {
-                let newElement = document.createElement(`${element.name}`)
-                newElement.setAttribute('id', `${element.id}`)
-                newElement.innerText = `this is a ${element.name}`
-                body.appendChild(newElement)
-            }
-        }
-    }, [elements])
-
 
     return (
         <div>
@@ -47,7 +27,6 @@ const ElementCreator = () => {
             <input type='text' placeholder='Enter element tag' onChange={handleChange}/>
             <button type='submit'>Add Element</button>
             </form>
-            
         </div>
     )
 }
